+++ src/main/java/bro/tasks/Deadline.java
@@ -1,29 +1,25 @@
+package bro.tasks;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-    protected LocalDateTime from;
-    protected LocalDateTime to;
+public class Deadline extends Task {
+    protected LocalDateTime by;
     private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
 
-    public Event(String description, String from, String to) {
+    public Deadline(String description, String by) {
         super(description);
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
         try {
-            this.from = LocalDateTime.parse(from, formatter);
-            this.to = LocalDateTime.parse(to, formatter);
+            this.by = LocalDateTime.parse(by, formatter);
         } catch (Exception e) {
             System.out.println("\tError parsing date/time. Please use the format: d/M/yyyy HHmm");
         }
     }
 
-    public Event(String description, boolean isDone, String from, String to) {
+    public Deadline(String description, boolean isDone, String by) {
         super(description, isDone);
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
         try {
-            this.from = LocalDateTime.parse(from, formatter);
-            this.to = LocalDateTime.parse(to, formatter);
+            this.by = LocalDateTime.parse(by, formatter);
         } catch (Exception e) {
             System.out.println("\tError parsing date/time. Please use the format: d/M/yyyy HHmm");
         }
@@ -32,13 +28,8 @@ public class Event extends Task {
     public boolean isOnDate(String date) {
         DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("d/M/yyyy");
         try {
-            LocalDate fromDate = from.toLocalDate();
-            LocalDate toDate = to.toLocalDate();
             LocalDate checkDate = LocalDate.parse(date, dateFormatter);
-            return (fromDate.isEqual(checkDate) 
-                || toDate.isEqual(checkDate)
-                || (checkDate.isAfter(fromDate) 
-                    && checkDate.isBefore(toDate)));
+            return by.toLocalDate().isEqual(checkDate);        
         } catch (Exception e) {
             System.out.println("\tError parsing date. Please use the format: d/M/yyyy");
             return false;
@@ -46,22 +37,20 @@ public class Event extends Task {
     }
 
     public String toDataString() {
-        return String.format("E|%d|%s|%s|%s", 
+        return String.format("D|%d|%s|%s", 
             (isDone ? 1 : 0), 
             description,
-            from.format(formatter),
-            to.format(formatter));
+            by.format(formatter));
     }
     
     @Override
     public String toString() {
         DateTimeFormatter printFormat = DateTimeFormatter.ofPattern("d MMM yyyy hh:mma");
         try {
-            return String.format("[E] [%s] %s (from: %s to: %s)", 
+            return String.format("[D] [%s] %s (by: %s)", 
                 getStatusIcon(), 
                 description,
-                from.format(printFormat),
-                to.format(printFormat));
+                by.format(printFormat));
         } catch (Exception e) {
             System.out.println("\tError parsing date. Please use the format: d/M/yyyy");
             return "";
