+++ src/main/java/bro/tasks/Task.java
@@ -1,3 +1,4 @@
+package bro.tasks;
 public class Task {
     protected String description;
     protected boolean isDone;
